push

+++ includes/addPage.inc.php
@@ -19,24 +19,28 @@
   */
   function addPage(string $content, string $name, string $code, int $authorID, string $date, int $online):bool{
     global $U, $USOC;
-    // Checks if the name already exists
-    $sql = "SELECT * FROM ". $U->contentHandlers[$content]["Name"] . " WHERE name='" . strtolower($name) . "';";
-    if(mysqli_num_rows(mysqli_query($U->db_link, $sql)) > 0){
+    if($U->contentHandlers[$content]){
+      // Checks if the name already exists
+      $sql = "SELECT * FROM ". $U->contentHandlers[$content]["Name"] . " WHERE name='" . strtolower($name) . "';";
+      if(mysqli_num_rows(mysqli_query($U->db_link, $sql)) > 0){
+        return False;
+      }
+      $sql = "INSERT INTO " . $U->contentHandlers[$content]["Name"] . " (Name, Code, Author, Date, Online) VALUES ('" . $name . "','" . addslashes($code) . "','" . $authorID . "','" . $date . "','" . $online . "');";
+      $db_erg = mysqli_query($U->db_link, $sql);
+      $sql = "SELECT * FROM ". $U->contentHandlers[$content]["Name"] . " WHERE name='" . $name . "';";
+      $db_erg2 = mysqli_query($U->db_link, $sql);
+      if($row = mysqli_fetch_array($db_erg2, MYSQLI_ASSOC)){
+        $id = $row["ID"];
+      }
+      if($U->contentHandlers[$content]["AddHandler"]($id, ["Name" => $name, "Code" => $code, "Author" => $authorID, "Date" => $date, "Online" => $online, "Id" => $id]) === False){
+        // Deletes the page if the AddHandler function returns false
+        $sql = "DELETE FROM ". $U->contentHandlers[$content]["Name"] . " WHERE name='" . $name . "';";
+        mysqli_query($U->db_link, $sql);
+        return False;
+      }
+      return $db_erg;
+    }else{
       return False;
     }
-    $sql = "INSERT INTO " . $U->contentHandlers[$content]["Name"] . " (Name, Code, Author, Date, Online) VALUES ('" . $name . "','" . addslashes($code) . "','" . $authorID . "','" . $date . "','" . $online . "');";
-    $db_erg = mysqli_query($U->db_link, $sql);
-    $sql = "SELECT * FROM ". $U->contentHandlers[$content]["Name"] . " WHERE name='" . $name . "';";
-    $db_erg2 = mysqli_query($U->db_link, $sql);
-    if($row = mysqli_fetch_array($db_erg2, MYSQLI_ASSOC)){
-      $id = $row["ID"];
-    }
-    if($U->contentHandlers[$content]["AddHandler"]($id, ["Name" => $name, "Code" => $code, "Author" => $authorID, "Date" => $date, "Online" => $online, "Id" => $id]) === False){
-      // Deletes the page if the AddHandler function returns false
-      $sql = "DELETE FROM ". $U->contentHandlers[$content]["Name"] . " WHERE name='" . $name . "';";
-      mysqli_query($U->db_link, $sql);
-      return False;
-    }
-    return $db_erg;
   }
 ?>
